migrate create cmd
